fix: rename secu priv
