+++ app/src/main/java/com/github/se/assocify/ui/screens/profile/ProfileGraph.kt
@@ -7,7 +7,7 @@ import com.github.se.assocify.model.database.UserAPI
 import com.github.se.assocify.navigation.Destination
 import com.github.se.assocify.navigation.NavigationActions
 import com.github.se.assocify.ui.screens.profile.notifications.profileNotificationsGraph
-import com.github.se.assocify.ui.screens.profile.security_privacy.profileSecurityPrivacyGraph
+import com.github.se.assocify.ui.screens.profile.securityPrivacy.profileSecurityPrivacyGraph
 import com.github.se.assocify.ui.screens.profile.theme.profileThemeGraph
 
 fun NavGraphBuilder.profileGraph(
